refactor(user-gateway): routes.py 를 책임별 sub-router 로 분할
기존 routes.py (259 LOC, 7 endpoint) 를 routes/ 패키지로 분리. SRP +
도메인별 가독성.

분할 결과:
- routes/__init__.py: 통합 router 조립 (main.py 호환)
- routes/health.py: healthz + Primary AgentCard 프록시 (system / proxy)
- routes/chat.py: POST /api/chat + GET /api/stream + ChatRequest (wire chat)
- routes/sessions.py: POST/GET/PATCH /api/sessions (session lifecycle / metadata)
- routes/history.py: GET /api/history (chats read)

main.py 의 `from user_gateway.routes import router` 는 그대로 호환.
route catalog (7 endpoint) 분할 전후 동일 검증.
